move ViolationHandler to common
