+++ verilog/analysis/BUILD
@@ -227,14 +227,13 @@ cc_library(
         "//common/analysis:text_structure_linter",
         "//common/analysis:token_stream_lint_rule",
         "//common/analysis:token_stream_linter",
-        "//common/strings:diff",
+        "//common/analysis:violation_handler",
         "//common/strings:line_column_map",
         "//common/text:concrete_syntax_tree",
         "//common/text:text_structure",
         "//common/text:token_info",
         "//common/util:file_util",
         "//common/util:logging",
-        "//common/util:user_interaction",
         "//verilog/parser:verilog_token_classifications",
         "//verilog/parser:verilog_token_enum",
         "@com_google_absl//absl/flags:flag",
@@ -307,6 +306,7 @@ cc_test(
         ":verilog_analyzer",
         ":verilog_linter",
         ":verilog_linter_configuration",
+        "//common/analysis:violation_handler",
         "//common/util:file_util",
         "//common/util:logging",
         "@com_google_absl//absl/memory",
