+++ verilog/analysis/verilog_linter.h
@@ -28,6 +28,7 @@
 #include "common/analysis/syntax_tree_linter.h"
 #include "common/analysis/text_structure_linter.h"
 #include "common/analysis/token_stream_linter.h"
+#include "common/analysis/violation_handler.h"
 #include "common/strings/line_column_map.h"
 #include "common/text/text_structure.h"
 #include "verilog/analysis/lint_rule_registry.h"
@@ -35,41 +36,10 @@
 
 namespace verilog {
 
-struct LintViolationWithStatus {
-  const verible::LintViolation* violation;
-  const verible::LintRuleStatus* status;
-
-  LintViolationWithStatus(const verible::LintViolation* v,
-                          const verible::LintRuleStatus* s)
-      : violation(v), status(s) {}
-
-  bool operator<(const LintViolationWithStatus& r) const {
-    // compares addresses which correspond to locations within the same string
-    return violation->token.text().data() < r.violation->token.text().data();
-  }
-};
-
 // Returns violations from multiple `LintRuleStatus`es sorted by position
 // of their occurrence in source code.
-std::set<LintViolationWithStatus> GetSortedViolations(
+std::set<verible::LintViolationWithStatus> GetSortedViolations(
     const std::vector<verible::LintRuleStatus>& statuses);
-
-// Interface for implementing violation handlers.
-//
-// The linting process produces a list of violations found in source code. Those
-// violations are then sorted and passed to `HandleViolations()` method of an
-// instance passed to LintOneFile().
-class ViolationHandler {
- public:
-  virtual ~ViolationHandler() = default;
-
-  // This method is called with a list of sorted violations found in file
-  // located at `path`. It can be called multiple times with statuses generated
-  // from different files. `base` contains source code from the file.
-  virtual void HandleViolations(
-      const std::set<LintViolationWithStatus>& violations,
-      absl::string_view base, absl::string_view path) = 0;
-};
 
 // Checks a single file for Verilog style lint violations.
 // This is suitable for calling from main().
@@ -86,7 +56,7 @@ class ViolationHandler {
 // errors were found (syntax, lint), and anything else is a fatal error.
 int LintOneFile(std::ostream* stream, absl::string_view filename,
                 const LinterConfiguration& config,
-                ViolationHandler* violation_handler, bool check_syntax,
+                verible::ViolationHandler* violation_handler, bool check_syntax,
                 bool parse_fatal, bool lint_fatal, bool show_context = false);
 
 // VerilogLinter analyzes a TextStructureView of Verilog source code.
@@ -133,108 +103,6 @@ absl::StatusOr<LinterConfiguration> LinterConfigurationFromFlags(
 // Expands linter configuration from a text file
 absl::Status AppendLinterConfigurationFromFile(
     LinterConfiguration* config, absl::string_view config_filename);
-
-// ViolationHandler that prints all violations in a form of user-friendly
-// messages.
-class ViolationPrinter : public ViolationHandler {
- public:
-  explicit ViolationPrinter(std::ostream* stream)
-      : stream_(stream), formatter_(nullptr) {}
-
-  void HandleViolations(const std::set<LintViolationWithStatus>& violations,
-                        absl::string_view base, absl::string_view path) final;
-
- protected:
-  std::ostream* const stream_;
-  verible::LintStatusFormatter* formatter_;
-};
-
-// ViolationHandler that prints all violations and gives an option to fix those
-// that have autofixes available.
-//
-// By default, when violation has an autofix available, ViolationFixer asks an
-// user what to do. The answers can be provided by AnswerChooser callback passed
-// to the constructor as the answer_chooser parameter. The callback is called
-// once for each fixable violation with a current violation object and a
-// violated rule name as arguments, and must return one of the values from
-// AnswerChoice enum.
-//
-// When the constructor's patch_stream parameter is not null, the fixes are
-// written to specified stream in unified diff format. Otherwise the fixes are
-// applied directly to the source file.
-//
-// The HandleLintRuleStatuses method can be called multiple times with statuses
-// generated from different files. The state of answers like "apply all for
-// rule" or "apply all" is kept between the calls.
-class ViolationFixer : public ViolationHandler {
- public:
-  enum class AnswerChoice {
-    kUnknown,
-    kApply,              // apply fix
-    kReject,             // reject fix
-    kApplyAllForRule,    // apply this and all remaining fixes for violations
-                         // of this rule
-    kRejectAllForRule,   // reject this and all remaining fixes for violations
-                         // of this rule
-    kApplyAll,           // apply this and all remaining fixes
-    kRejectAll,          // reject this and all remaining fixes
-    kPrintFix,           // show fix
-    kPrintAppliedFixes,  // show fixes applied so far
-  };
-
-  struct Answer {
-    AnswerChoice choice;
-    // If there are multiple alternatives for fixes available, this is
-    // the one chosen. By default the first one.
-    size_t alternative = 0;
-  };
-
-  using AnswerChooser =
-      std::function<Answer(const verible::LintViolation&, absl::string_view)>;
-
-  // Violation fixer with user-chosen answer chooser.
-  ViolationFixer(std::ostream* message_stream, std::ostream* patch_stream,
-                 const AnswerChooser& answer_chooser)
-      : ViolationFixer(message_stream, patch_stream, answer_chooser, false) {}
-
-  // Violation fixer with interactive answer choice.
-  ViolationFixer(std::ostream* message_stream, std::ostream* patch_stream)
-      : ViolationFixer(message_stream, patch_stream, InteractiveAnswerChooser,
-                       true) {}
-
-  void HandleViolations(const std::set<LintViolationWithStatus>& violations,
-                        absl::string_view base, absl::string_view path) final;
-
- private:
-  ViolationFixer(std::ostream* message_stream, std::ostream* patch_stream,
-                 const AnswerChooser& answer_chooser, bool is_interactive)
-      : message_stream_(message_stream),
-        patch_stream_(patch_stream),
-        answer_chooser_(answer_chooser),
-        is_interactive_(is_interactive),
-        ultimate_answer_({AnswerChoice::kUnknown, 0}) {}
-
-  void HandleViolation(const verible::LintViolation& violation,
-                       absl::string_view base, absl::string_view path,
-                       absl::string_view url, absl::string_view rule_name,
-                       const verible::LintStatusFormatter& formatter,
-                       verible::AutoFix* fix);
-
-  static Answer InteractiveAnswerChooser(
-      const verible::LintViolation& violation, absl::string_view rule_name);
-
-  void CommitFixes(absl::string_view source_content,
-                   absl::string_view source_path,
-                   const verible::AutoFix& fix) const;
-
-  std::ostream* const message_stream_;
-  std::ostream* const patch_stream_;
-  const AnswerChooser answer_chooser_;
-  const bool is_interactive_;
-
-  Answer ultimate_answer_;
-  std::map<absl::string_view, Answer> rule_answers_;
-};
 
 // VerilogLintTextStructure analyzes Verilog syntax tree for style violations
 // and syntactically detectable pitfalls.
