+++ verilog/tools/lint/BUILD
@@ -230,6 +230,7 @@ cc_binary(
         "//common/util:file_util",
         "//common/util:init_command_line",
         "//common/util:logging",
+        "//common/analysis:violation_handler",
         "//verilog/analysis:verilog_linter",
         "//verilog/analysis:verilog_linter_configuration",
         "@com_google_absl//absl/flags:flag",
